+++ src/main/java/gr/james/sampling/EfraimidisSampling.java
@@ -17,7 +17,7 @@ public class EfraimidisSampling<T> implements WeightedRandomSampling<T> {
     private final int sampleSize;
     private final Random random;
     private final PriorityQueue<Weighted<T>> pq;
-    private int streamSize;
+    private long streamSize;
 
     /**
      * Construct a new instance of {@link EfraimidisSampling} using the specified sample size and RNG. The
@@ -57,7 +57,7 @@ public class EfraimidisSampling<T> implements WeightedRandomSampling<T> {
         if (item == null) {
             throw new NullPointerException("Item was null");
         }
-        if (streamSize == Integer.MAX_VALUE) {
+        if (streamSize == Long.MAX_VALUE) {
             throw new StreamOverflowException();
         }
         if (weight <= 0) {
@@ -125,7 +125,7 @@ public class EfraimidisSampling<T> implements WeightedRandomSampling<T> {
      * This method runs in constant time.
      */
     @Override
-    public final int streamSize() {
+    public final long streamSize() {
         assert this.streamSize >= 0;
         return this.streamSize;
     }
